feat: create mapper for ThContract
